Update of PSim based ChoiceSetGenerator
git-svn-id: file:///srv/local/data/cleanRepo/trunk@32870 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/artemc/src/main/java/playground/artemc/heterogeneity/IncomeHeterogeneityModule.java
@@ -80,7 +80,7 @@ public class IncomeHeterogeneityModule extends AbstractModule {
 				log.info("Simulation with "+incomeHeterogeneityImpl.getType()+" heterogeneity: calculating income factors.");
 
 				/*Calculate Income Statistics*/
-				Integer incomeSum=0;
+				Long incomeSum= Long.valueOf(0);
 				Double incomeMean = 0.0;
 
 				for(Id<Person> personId:population.getPersons().keySet()){
